ENG-543 decoupling contents table filter + selector to a new component
